pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "balkan-nlp"
version = "0.1.0"
description = "High-quality NLP datasets & pipelines for Serbian, Bosnian, and Croatian (sr/bs/hr)"
readme = "README.md"
requires-python = ">=3.11"
license = { text = "MIT" }
authors = [
  { name = "Balkan NLP Project", email = "office@rsateam.com" }
]
keywords = ["nlp", "datasets", "serbian", "bosnian", "croatian", "sr", "bs", "hr", "balkans"]
classifiers = [
  "Development Status :: 3 - Alpha",
  "Intended Audience :: Science/Research",
  "Intended Audience :: Developers",
  "Topic :: Scientific/Engineering :: Artificial Intelligence",
  "Topic :: Text Processing :: Linguistic",
  "License :: OSI Approved :: MIT License",
  "Programming Language :: Python :: 3",
  "Programming Language :: Python :: 3.11",
  "Programming Language :: Python :: 3.12",
  "Programming Language :: Python :: 3.13",
]

dependencies = [
  "requests>=2.32.5",
  "beautifulsoup4>=4.14.3",
  "trafilatura>=2.0.0",
  "datasets>=3.3.2",
  "pandas>=2.3.3",
  "numpy>=2.4.1",
  "tqdm>=4.67.1",
  "pyyaml>=6.0.3",
  "python-dateutil>=2.9.0.post0",
  "typer>=0.21.1",
  "mwparserfromhell>=0.7.2",
  "mwxml>=0.3.6",
  "feedparser>=6.0.12",
]

[project.optional-dependencies]
dev = [
  "pytest>=9.0.2",
  "black>=26.1.0",
  "isort>=7.0.0",
  "flake8>=7.3.0",
  "mypy>=1.19.1",
  "pre-commit>=4.5.1",
]

processing = [
  "spacy>=3.8.11",
  "langdetect>=1.0.9",
  "datasketch>=1.9.0",
  "fasttext>=0.9.3",
]

export = [
  "pyarrow>=23.0.0",
  "huggingface-hub>=1.3.2",
]

all = [
  "balkan-nlp[dev,processing,export]",
]

[project.urls]
Homepage = "https://github.com/rsadevteam/balkan-nlp"
Documentation = "https://github.com/rsadevteam/balkan-nlp/blob/main/README.md"
Repository = "https://github.com/rsadevteam/balkan-nlp"
Issues = "https://github.com/rsadevteam/balkan-nlp/issues"

[tool.setuptools]
package-dir = { "" = "." }

[tool.setuptools.packages.find]
where = ["."]
include = ["scraping*", "processing*", "export*", "utils*"]
exclude = ["tests*", "docs*", "scripts*"]

[tool.black]
line-length = 100
target-version = ["py311", "py312", "py313"]
include = '\.pyi?$'
exclude = '''
/(
    \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | _build
  | buck-out
  | build
  | dist
)/
'''

[tool.isort]
profile = "black"
line_length = 100
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
disallow_incomplete_defs = false
check_untyped_defs = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
follow_imports = "normal"
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v --tb=short"
